+++ .gitignore
@@ -1,0 +1,3 @@
+token.txt
+/.idea/*
+/__pycache__/*
