+++ README.md
@@ -1,3 +1,1 @@
 # Team_8
-
-Тема проекта  - сделать бота-визуализатора, который по заданным данным нарисует график или диаграмму а также исследует заданную функцию.
