+++ main.py
@@ -1,11 +1,11 @@
-import classBot
+import bot
 
 
 def main():
-    token = ''  # for_token
-    b = classBot.Bot(token)
-    while True:
-        b.work()
+    f = open('token.txt')  # create token.txt wint your token
+    token = f.read()
+    b = bot.Bot(token)
+    b.work()
 
 
 if __name__ == '__main__':
